分离LogView组件
